refactor api executor
